Fix program-crippling bug that effectively made the -r param the same as -m
This fixes a massive, stupid bug, in existence for this utility's entire lifetime, which effectively limited the -r parameter to the same value as the -m parameter. When computing the size of the Opcode check window, detectOpcodeAddition() mistakenly added seqLen (the -m parameter) when it was intended to add maxDist (the -r parameter). Since seqLen is typically set to a much lower value than maxDist in order to be useful, this massively reduced the usefulness of the program.

It's incredibly dumb, but at least I figured it out after...oh, four years or so of wondering why I kept getting wacky results no matter what parameters I used!